bot/requirements.txt
aiogram>=3.4,<4.0
aiohttp>=3.9
python-dotenv>=1.0

